gandia-huella/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
opencv-python==4.9.0.80
opencv-contrib-python==4.9.0.80
torch
torchvision
scikit-learn
numpy
requests
Pillow
python-dotenv==1.0.1
scipy
